chatbot/python/requirements.txt
Pillow
cairosvg
numpy
spidev
opencv-python-headless
Flask
requests
RPi.GPIO